JobFlow-AI/server/requirements.txt
# ============================================================
# JobFlow-AI — Python Dependencies
# ============================================================

# Backend API
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.10.0
pydantic-settings==2.6.0
python-dotenv==1.0.1

# Database
sqlalchemy==2.0.36
alembic==1.14.0
psycopg2-binary==2.9.10
aiosqlite==0.20.0

# Scraping
playwright==1.49.0
beautifulsoup4==4.12.3
lxml==5.3.0

# OSINT
httpx==0.28.0
ddgs>=9.0.0
dnspython>=2.7.0

# AI
google-generativeai==0.8.3

# PDF Generation
weasyprint==63.0
jinja2==3.1.4

# Email
python-slugify==8.0.4

# Task Queue
celery[redis]==5.4.0
redis==5.2.0

# Utilities
python-multipart==0.0.12
aiofiles==24.1.0
